+++ documentation/CHANGELOG.md
@@ -12,8 +12,9 @@
 # Releases
 <!-- @LatestFirst -->
 
-## [0.3.1] [![user](https://img.shields.io/badge/jrsmth-181717.svg?style=flat&logo=github)](https://github.com/jrsmth)
-[Scroll Command](https://github.com/jrsmth/waffle-bot/milestone/4) (08/04/2024)
+## [0.3.1] [![user](https://img.shields.io/badge/adamj335-181717.svg?style=flat&logo=github)](https://github.com/adamj335) [![user](https://img.shields.io/badge/jrsmth-181717.svg?style=flat&logo=github)](https://github.com/jrsmth)
+[Scroll Command](https://github.com/jrsmth/waffle-bot/milestone/4) (11/04/2024)
+- `#44` Prevent duplicate records for the same streak
 - `#46` Add README Badges (Code Coverage, Workflow, Deployment)
 
 ## [0.3.0] [![user](https://img.shields.io/badge/adamj335-181717.svg?style=flat&logo=github)](https://github.com/adamj335)